+++ src/scene.rs
@@ -74,7 +74,7 @@ pub fn replicate_into(scene: &mut DynamicScene, world: &World) {
         }
 
         for rule in rules.iter().filter(|rule| rule.matches(archetype)) {
-            for &(component_id, _) in &rule.components {
+            for &(component_id, _) in rule.components() {
                 // SAFETY: replication rules can be registered only with valid component IDs.
                 let replicated_component =
                     unsafe { world.components().get_info_unchecked(component_id) };