+++ src/server/replicated_archetypes.rs
@@ -47,7 +47,7 @@ impl ReplicatedArchetypes {
         {
             let mut replicated_archetype = ReplicatedArchetype::new(archetype.id());
             for rule in rules.iter().filter(|rule| rule.matches(archetype)) {
-                for &(component_id, fns_id) in &rule.components {
+                for &(component_id, fns_id) in rule.components() {
                     // Since rules are sorted by priority,
                     // we are inserting only new components that aren't present.
                     if replicated_archetype